+++ packages/happn-commons/lib/constants-builder.js
@@ -1,6 +1,6 @@
 const constants = require('./constants');
 
-Object.keys(constants).forEach(function(constantBranch) {
+Object.keys(constants).forEach(function (constantBranch) {
   constants[`${constantBranch}_COLLECTION`] = Object.values(constants[constantBranch]);
 });