+++ packages/happn-db-provider-mongo/RELEASES.md
@@ -123,3 +123,7 @@
 -----------------
   - fix: happn-3 resetpassword
   - feat: happner-2 resetpassword
+
+1.1.11 2023-08-05
+-----------------
+- feat: fixed mesh_gst paths in happner-2