+++ packages/happn-db-provider-sqlite/RELEASES.md
@@ -34,3 +34,7 @@
 -----------------
   - fix: happn-3 resetpassword
   - feat: happner-2 resetpassword
+
+1.0.8 2023-08-05
+-----------------
+- feat: fixed mesh_gst paths in happner-2